Relocatable ocamlbuild
In configure.make, specifying a path beginning ./ or ../ or
OCAMLBUILD_LIBDIR enabled Relocatable ocamlbuild, as long as the
ocamlbuild binary is being installed in the same directory as the
compiler and the compiler itself is Relocatable OCaml. The relative path
given for OCAMLBUILD_LIBDIR describes how to get from OCaml's Standard
Library to the directory containing the "ocamlbuild" package directory
(in opam, this is just "..").

In this mode, `ocamlbuild -where` is determined relative to the Standard
Library.